Clean up old sqlboiler repository implementations
Remove all old sqlboiler repository implementations and replace with
interface definitions. The sqlc implementations (created in Phase 5)
are now the only implementations of these interfaces.

Changes:
- Delete 13 old sqlboiler repository files (8 watchrepo + 5 coldrepo)
- Delete 2 old test files that used sqlboiler repositories
- Create pkg/repository/coldrepo/interfaces.go with 5 interfaces
- Create pkg/repository/watchrepo/interfaces.go with 8 interfaces
- Remove old constructor functions from pkg/testutil/repository.go
- All three wallets (watch, keygen, sign) compile successfully

This completes the sqlboiler to sqlc migration cleanup.
