+++ pkg/repository/coldrepo/interfaces.go
@@ -1,0 +1,58 @@
+package coldrepo
+
+import (
+	"github.com/hiromaily/go-crypto-wallet/pkg/account"
+	"github.com/hiromaily/go-crypto-wallet/pkg/address"
+	models "github.com/hiromaily/go-crypto-wallet/pkg/models/rdb"
+)
+
+// SeedRepositorier is SeedRepository interface
+type SeedRepositorier interface {
+	GetOne() (*models.Seed, error)
+	Insert(strSeed string) error
+}
+
+// AccountKeyRepositorier is AccountKeyRepository interface
+type AccountKeyRepositorier interface {
+	GetMaxIndex(accountType account.AccountType) (int64, error)
+	GetOneMaxID(accountType account.AccountType) (*models.AccountKey, error)
+	GetAllAddrStatus(accountType account.AccountType, addrStatus address.AddrStatus) ([]*models.AccountKey, error)
+	GetAllMultiAddr(accountType account.AccountType, addrs []string) ([]*models.AccountKey, error)
+	InsertBulk(items []*models.AccountKey) error
+	UpdateAddr(accountType account.AccountType, addr, keyAddress string) (int64, error)
+	UpdateAddrStatus(accountType account.AccountType, addrStatus address.AddrStatus, strWIFs []string) (int64, error)
+	UpdateMultisigAddr(accountType account.AccountType, item *models.AccountKey) (int64, error)
+	UpdateMultisigAddrs(accountType account.AccountType, items []*models.AccountKey) (int64, error)
+}
+
+// XRPAccountKeyRepositorier is XRPAccountKeyRepository interface
+type XRPAccountKeyRepositorier interface {
+	GetAllAddrStatus(accountType account.AccountType, addrStatus address.AddrStatus) ([]*models.XRPAccountKey, error)
+	GetSecret(accountType account.AccountType, addr string) (string, error)
+	InsertBulk(items []*models.XRPAccountKey) error
+	UpdateAddrStatus(accountType account.AccountType, addrStatus address.AddrStatus, strWIFs []string) (int64, error)
+}
+
+// AuthFullPubkeyRepositorier is AuthFullPubkeyRepository interface
+type AuthFullPubkeyRepositorier interface {
+	GetOne(authType account.AuthType) (*models.AuthFullpubkey, error)
+	Insert(authType account.AuthType, fullPubKey string) error
+	InsertBulk(items []*models.AuthFullpubkey) error
+}
+
+// AuthAccountKeyRepositorier is AuthAccountKeyRepository interface
+type AuthAccountKeyRepositorier interface {
+	GetOne(authType account.AuthType) (*models.AuthAccountKey, error)
+	Insert(item *models.AuthAccountKey) error
+	UpdateAddrStatus(addrStatus address.AddrStatus, strWIF string) (int64, error)
+}
+
+// GetRedeedScriptByAddress returns redeem script by address
+func GetRedeedScriptByAddress(accountKeys []*models.AccountKey, addr string) string {
+	for _, val := range accountKeys {
+		if val.MultisigAddress == addr {
+			return val.RedeemScript
+		}
+	}
+	return ""
+}
